# my-app/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "Shopping Service API Specialized in coffee",
    "version": "1.0",
    "title": "bean cafe API",
    "termsOfService": "about:blank",
    "contact": {
      "name": "Hongrok Lim",
      "url": "https://github.com/rokong/",
      "email": "hongrr123@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8080",
  "basePath": "/bean-cafe",
...
  "paths": {
...
    "/order/{orderId}/product": {
      "get": {
        "tags": [
          "Order Product"
        ],
        "summary": "get order products",
        "description": "get products by order id [orderId]",
        "operationId": "getOProductsUsingGET",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "description": "orderId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Order Product"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "Order Product"
        ],
        "summary": "add order product",
        "description": "add product in order [orderId, productId, (optionCd), cnt, (payId)]",
        "operationId": "addOProductUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "oProduct",
            "description": "oProduct",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Order Product"
            }
          },
          {
            "name": "orderId",
            "in": "path",
            "description": "orderId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Order Product"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/order/{orderId}/product/{productId}/{optionCd}": {
      "get": {
        "tags": [
          "Order Product"
        ],
        "summary": "get order product",
        "description": "get order product [orderId, productId, (optionCd)]",
        "operationId": "getOProductUsingGET",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "optionCd",
            "in": "path",
            "description": "optionCd",
            "required": true,
            "type": "string"
          },
          {
            "name": "orderId",
            "in": "path",
            "description": "orderId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "productId",
            "in": "path",
            "description": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Order Product"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "Order Product"
        ],
        "summary": "update count",
        "description": "update order product's count [orderId, productId, (optionCd), cnt]",
        "operationId": "updateOProductUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "oProduct",
            "description": "oProduct",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Order Product"
            }
          },
          {
            "name": "optionCd",
            "in": "path",
            "description": "optionCd",
            "required": true,
            "type": "string"
          },
          {
            "name": "orderId",
            "in": "path",
            "description": "orderId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "productId",
            "in": "path",
            "description": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Order Product"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "Order Product"
        ],
        "summary": "delete order product",
        "description": "delete order product [orderId, productId, (optionCd)]",
        "operationId": "deleteOProductUsingDELETE",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "optionCd",
            "in": "path",
            "description": "optionCd",
            "required": true,
            "type": "string"
          },
          {
            "name": "orderId",
            "in": "path",
            "description": "orderId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "productId",
            "in": "path",
            "description": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
    "/product": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "get product list",
        "description": "get product list",
        "operationId": "getProductListUsingGET",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "create product",
        "description": "create product [name, price, (categoryId), (enabled),sellerNm, (stockCnt), (deliveryId), (deliveryPrice), (discountPrice)]",
        "operationId": "createProductUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "product",
            "description": "product",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/product/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "get product",
        "description": "get product by id",
        "operationId": "getProductUsingGET",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "Product"
        ],
        "summary": "update product",
        "description": "[(name), (price), (categoryId), (stockCnt), (enabled),(deliveryId), (discountPrice)]",
        "operationId": "updateProductUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "product",
            "description": "product",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "summary": "delete product",
        "description": "delete product",
        "operationId": "deleteProductUsingDELETE",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
    "/product/{productId}/option": {
      "get": {
        "tags": [
          "Product Option"
        ],
        "summary": "get options in product",
        "description": "get product options in product",
        "operationId": "getPOptionsInProductUsingGET",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product Option"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/product/{productId}/option/group": {
      "post": {
        "tags": [
          "Product Option"
        ],
        "summary": "create option group",
        "description": "create product option group [name]",
        "operationId": "createProductOptionGroupUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "pOption",
            "description": "pOption",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product Option"
            }
          },
          {
            "name": "productId",
            "in": "path",
            "description": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product Option"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "Product Option"
        ],
        "summary": "delete option group",
        "description": "delete last product option group",
        "operationId": "deleteProductOptionGroupUsingDELETE",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
    "/product/{productId}/option/{groupId}": {
      "get": {
        "tags": [
          "Product Option"
        ],
        "summary": "get options in group",
        "description": "get product options in product and group",
        "operationId": "getPOptionsInGroupUsingGET",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "groupId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "productId",
            "in": "path",
            "description": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product Option"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "Product Option"
        ],
        "summary": "create option",
        "description": "create product option []",
        "operationId": "createProductOptionUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "groupId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "pOption",
            "description": "pOption",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product Option"
            }
          },
          {
            "name": "productId",
            "in": "path",
            "description": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product Option"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "Product Option"
        ],
        "summary": "update option group's order",
        "description": "update product option group's order [ord]",
        "operationId": "updateProductOptionGroupOrderUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "groupId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "pOption",
            "description": "pOption",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product Option"
            }
          },
          {
            "name": "productId",
            "in": "path",
            "description": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product Option"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/product/{productId}/option/{groupId}/{optionId}": {
      "get": {
        "tags": [
          "Product Option"
        ],
        "summary": "get option",
        "description": "get product option",
        "operationId": "getPOptionUsingGET",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "groupId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "optionId",
            "in": "path",
            "description": "optionId",
            "required": true,
            "type": "string"
          },
          {
            "name": "productId",
            "in": "path",
            "description": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product Option"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "Product Option"
        ],
        "summary": "update option",
        "description": "update option [(name), (ord)]",
        "operationId": "updateProductOptionUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "groupId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "optionId",
            "in": "path",
            "description": "optionId",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "pOption",
            "description": "pOption",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Product Option"
            }
          },
          {
            "name": "productId",
            "in": "path",
            "description": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Product Option"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "Product Option"
        ],
        "summary": "delete option",
        "description": "delete product option",
        "operationId": "deleteProductOptionUsingDELETE",
        "produces": [
          "application/json;charset=UTF-8"
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "groupId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "optionId",
            "in": "path",
            "description": "optionId",
            "required": true,
            "type": "string"
          },
          {
            "name": "productId",
            "in": "path",
            "description": "productId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
...
    "Order": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32",
          "example": 1,
          "description": "order id"
        },
        "userNm": {
          "type": "string",
          "example": "customer1",
          "description": "customer name"
        },
        "price": {
          "type": "integer",
          "format": "int32",
          "example": 13700,
          "description": "products' price"
        },
        "deliveryPrice": {
          "type": "integer",
          "format": "int32",
          "example": 5500,
          "description": "delivery price"
        },
        "payId": {
          "type": "integer",
          "format": "int32",
          "example": 7,
          "description": "pay id"
        },
        "payNm": {
          "type": "string",
          "description": "pay name"
        },
        "requestDt": {
          "type": "string",
          "example": "2020-10-18",
          "description": "order request date"
        },
        "lastEditDt": {
          "type": "string",
          "example": "2020-10-18",
          "description": "last changed date"
        },
        "orderStatus": {
          "type": "string",
          "example": "WRITING",
          "description": "order status",
          "enum": [
            "WRITING",
            "PAYMENT",
            "PAYMENT_STANDBY",
            "PAYMENT_PROGRESS",
            "CHECKING",
            "CHECK_COMPLETE",
            "PRODUCT_READY",
            "DELIVERING",
            "COMPLETE",
            "CANCEL",
            "CUSTOMER_CANCEL",
            "CANCELED_WRITE",
            "CANCELED_PAYMENT",
            "CANCELED_CHECK",
            "CANCELED_PRODUCT",
            "SELLER_CANCEL",
            "CANCEL_CHECK",
            "CANCEL_PRODUCT",
            "CANCEL_DELIVER"
          ]
        }
      },
      "title": "Order",
      "description": "Order including primary information"
    },
...
    "Product": {
      "type": "object",
      "properties": {
        "stockCntInt": {
          "type": "integer",
          "format": "int32"
        },
        "id": {
          "type": "integer",
          "format": "int32",
          "example": 2,
          "description": "product id"
        },
        "name": {
          "type": "string",
          "example": "Bean Cafe Main Blending",
          "description": "product name"
        },
        "price": {
          "type": "integer",
          "format": "int32",
          "example": 8000,
          "description": "product price"
        },
        "categoryId": {
          "type": "integer",
          "format": "int32",
          "example": 3,
          "description": "product's category"
        },
        "enabled": {
          "type": "boolean",
          "example": true,
          "description": "is enabled"
        },
        "sellerNm": {
          "type": "string",
          "example": "seller1",
          "description": "seller name"
        },
        "stockCnt": {
          "type": "integer",
          "format": "int32",
          "description": "stock amount (can be empty)"
        },
        "deliveryId": {
          "type": "integer",
          "format": "int32",
          "example": 2,
          "description": "delivery id"
        },
        "deliveryPrice": {
          "type": "integer",
          "format": "int32",
          "example": 3000,
          "description": "delivery price"
        },
        "discountPrice": {
          "type": "integer",
          "format": "int32",
          "example": -800,
          "description": "discount price"
        }
      },
      "title": "Product",
      "description": "Product including primary information"
    },
...
  }
}